requirements.txt
APScheduler==3.7.0
bokeh==2.2.3
branca==0.4.2
CacheControl==0.12.6
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
cryptography==3.3.1
cycler==0.10.0
DateTime==4.3
decorator==4.4.2
emoji==0.6.0
firebase-admin==4.5.1
Flask==1.1.2
Flask-Executor==0.9.4
Flask-Login==0.5.0
Flask-WTF==0.14.3
folium==0.12.1
funcy==1.15
future==0.18.2
gcloud==0.18.3
gensim==3.8.3
geographiclib==1.50
geopy==2.1.0
gmplot==1.4.1
google-api-core==1.25.1
google-api-python-client==1.12.8
google-auth==1.24.0
google-auth-httplib2==0.0.4
google-cloud-core==1.5.0
google-cloud-firestore==2.0.2
google-cloud-storage==1.35.0
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.52.0
grpcio==1.35.0
httplib2==0.18.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==1.0.0
jwcrypto==0.8
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.3
msgpack==1.0.2
networkx==2.5
nltk==3.5
numexpr==2.7.2
numpy==1.19.5
oauth2client==4.1.3
oauthlib==3.1.0
packaging==20.8
pandas==1.2.0
pathlib==1.0.1
Pillow==8.1.0
proto-plus==1.13.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.9.9
pyLDAvis==3.2.2
pyparsing==2.4.7
pyproj==3.0.0.post1
Pyrebase4==4.4.3
pyshp==2.1.3
PySocks==1.7.1
python-dateutil==2.8.1
python-jwt==3.3.0
pytz==2020.5
PyYAML==5.4.1
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==4.7
schedule==0.6.0
scikit-learn==0.24.0
scipy==1.6.0
selenium==3.141.0
six==1.15.0
sklearn==0.0
smart-open==4.2.0
SQLAlchemy==1.3.22
tabulate==0.8.7
textblob==0.15.3
threadpoolctl==2.1.0
tornado==6.1
tqdm==4.56.0
tweepy==3.10.0
typing-extensions==3.7.4.3
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.26.2
Werkzeug==1.0.1
wordcloud==1.8.1
WTForms==2.3.3
zope.interface==5.2.0
